+++ posthog/api/organization.py
@@ -12,8 +12,7 @@ from rest_framework import exceptions, generics, permissions, response, serializ
 from rest_framework.request import Request
 
 from posthog.api.routing import StructuredViewSetMixin
-from posthog.api.shared import UserBasicSerializer
-from posthog.api.team import TeamBasicSerializer
+from posthog.api.shared import TeamBasicSerializer, UserBasicSerializer
 from posthog.demo import create_demo_team
 from posthog.event_usage import report_onboarding_completed, report_user_joined_organization, report_user_signed_up
 from posthog.mixins import AnalyticsDestroyModelMixin
@@ -60,20 +59,6 @@ class OrganizationPermissionsWithDelete(OrganizationAdminWritePermissions):
             OrganizationMembership.Level.OWNER if request.method == "DELETE" else OrganizationMembership.Level.ADMIN
         )
         return OrganizationMembership.objects.get(user=request.user, organization=organization).level >= min_level
-
-
-class OrganizationBasicSerializer(serializers.ModelSerializer):
-    """
-    Serializer for `Organization` model with minimal attributes to speeed up loading and transfer times.
-    Also used for nested serializers.
-    """
-
-    class Meta:
-        model = Organization
-        fields = [
-            "id",
-            "name",
-        ]
 
 
 class OrganizationSerializer(serializers.ModelSerializer):
